✅ test(favorite): 비결정적 동시성 테스트를 퍼포먼스 테스트로 분리
